+++ packages/app/pages/course/[cid]/queue/[qid].tsx
@@ -3,16 +3,13 @@ import {
   ClosedQuestionStatus,
   OpenQuestionStatus,
   Question,
-  QuestionStatus,
   QuestionType,
   Role,
 } from "@template/common";
 import { useRouter } from "next/router";
-import { useCallback, useState } from "react";
 import styled from "styled-components";
 import useSWR, { mutate } from "swr";
 import NavBar from "../../../../components/Nav/NavBar";
-import StudentPopupCard from "../../../../components/Queue/StudentPopupCard";
 import StudentQueueList from "../../../../components/Queue/StudentQueueList";
 import TAQueueList from "../../../../components/Queue/TAQueueList";
 import { useProfile } from "../../../../hooks/useProfile";
@@ -46,20 +43,6 @@ export default function Queue() {
   const studentQuestion =
     profile && questions && questions.find((q) => q.creator.id === profile.id);
 
-  // TA queue state variables
-  const [openPopup, setOpenPopup] = useState<boolean>(false);
-  const [currentQuestion, setCurrentQuestion] = useState<Question>(null);
-
-  const onOpenClick = useCallback((question: Question): void => {
-    setCurrentQuestion(question);
-    setOpenPopup(true);
-  }, []);
-
-  const onCloseClick = useCallback((): void => {
-    setCurrentQuestion(null);
-    setOpenPopup(false);
-  }, []);
-
   const joinQueue = async () => {
     const createdQuestion = await API.questions.create({
       queueId: Number(qid),
@@ -92,29 +75,6 @@ export default function Queue() {
     mutate(`/api/v1/queues/${qid}/questions`, newQuestions);
   };
 
-  /**
-   * TA functions to support queue operations
-   */
-
-  /**
-   * Updates a given question to the given status.
-   * @param question the question being modified
-   * @param status the updated status
-   */
-  const updateQuestionTA = async (
-    question: Question,
-    status: QuestionStatus
-  ) => {
-    await API.questions.update(question.id, {
-      status: status,
-    });
-    const newQuestions = questions.map((q) =>
-      q.id === question.id ? { ...q, status } : q
-    );
-    mutate(`/api/v1/queues/${qid}/questions`, newQuestions);
-    setOpenPopup(false);
-  };
-
   const alertStudent = async (question: Question) => {
     await API.questions.notify(question.id);
   };
@@ -128,10 +88,8 @@ export default function Queue() {
             {Role.STUDENT === role ? (
               <StudentQueueList
                 room={""}
-                onOpenClick={onOpenClick}
                 joinQueue={joinQueue}
                 questions={questions}
-                helpingQuestions={helpingQuestions}
                 studentQuestion={studentQuestion}
                 leaveQueue={leaveQueue}
                 finishQuestion={finishQuestion}
@@ -139,24 +97,11 @@ export default function Queue() {
             ) : (
               <TAQueueList
                 qid={Number(qid)}
-                onOpenClick={onOpenClick}
-                updateQuestionTA={updateQuestionTA}
                 alertStudent={alertStudent}
                 questions={questions}
                 helpingQuestions={helpingQuestions}
                 groupQuestions={groupQuestions}
                 courseId={Number(cid)}
-              />
-            )}
-            {role === Role.TA && currentQuestion && (
-              <StudentPopupCard
-                onClose={onCloseClick}
-                email="takayama.a@northeastern.edu" //need a way to access this. or the user
-                wait={20} //figure out later
-                question={currentQuestion}
-                location="Outside by the printer" // need a way to access this
-                visible={openPopup}
-                updateQuestion={updateQuestionTA}
               />
             )}
           </>
