+++ packages/app/pages/login.tsx
@@ -56,7 +56,17 @@ export default function Login({}: LoginProps) {
             htmlType="submit"
             onClick={onSubmit}
           >
-            Login as <b>Student</b>
+            Login as <b>Student 1</b>
+          </Button>
+        </Link>
+        <Link href="/api/v1/profile/entry?userId=3">
+          <Button
+            style={{ marginRight: "15px" }}
+            type="default"
+            htmlType="submit"
+            onClick={onSubmit}
+          >
+            Login as <b>Student 2</b>
           </Button>
         </Link>
         <Link href="/api/v1/profile/entry?userId=2">
